backend/requirements.txt
# FastAPI 관련
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 데이터베이스
sqlalchemy==2.0.28
alembic==1.13.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# 데이터 검증
pydantic==2.6.3
pydantic[email]==2.6.3
email-validator==2.1.1

# 기타
requests==2.31.0
python-dateutil==2.9.0